add,delete course

+++ server/controllers/admin.js
@@ -2,6 +2,9 @@ import TryCatch from "../middlewares/TryCatch.js";
 import { Courses } from "../models/Courses.js";
 import { Lecture } from "../models/Lecture.js";
 import { rm } from "fs";
+import { promisify } from "util";
+import fs from "fs";
+import { User } from "../models/User.js";
 
 export const createCourse = TryCatch(async (req, res) => {
   const { title, description, category, createdBy, duration, price } = req.body;
@@ -53,9 +56,41 @@ export const addlectures = TryCatch(async (req, res) => {
 
 export const deleteLecture = TryCatch(async (req, res) => {
   const lecture = await Lecture.findById(req.params.id);
+
+  if (!lecture)
+    return res.status(404).json({ message: "Lecture not found" });
+
   rm(lecture.video, () => {
     console.log("Video Deleted");
   });
   await lecture.deleteOne();
   res.json({ message: "Lecture Deleted" });
+});
+
+const unlinkAsync = promisify(fs.unlink);
+
+export const deleteCourse = TryCatch(async (req, res) => {
+  const course = await Courses.findById(req.params.id);
+
+  const lectures = await Lecture.find({ course: course._id });
+
+  await Promise.all(
+    lectures.map(async (lecture) => {
+      await unlinkAsync(lecture.video);
+      console.log("Video Deleted");
+    }),
+  );
+  rm(course.image, () => {
+    console.log("image Deleted");
+  });
+
+  await Lecture.find({ course: req.params.id }).deleteMany();
+
+  await course.deleteOne();
+
+  await User.updateMany({}, { $pull: { subscription: req.params.id } });
+
+  res.json({
+    message: "Course Deleted",
+  });
 });
